Update to Develop from Patch-3

+++ App/Settings.cs
@@ -17,7 +17,7 @@ namespace App
         public static bool AutoUpdate { get; set; } = true;
         public static bool TwitterEnabled { get; set; } = false;
         public static string TwitterAccount { get; set; } = "";
-        public static HashSet<ushort> FATEs { get; set; } = new HashSet<ushort>();
+        public static HashSet<int> FATEs { get; set; } = new HashSet<int>();
 
         private static void Init()
         {
@@ -48,7 +48,7 @@ namespace App
                 string fates = iniFile.ReadValue("fate", "fates");
                 if (!string.IsNullOrEmpty(fates))
                 {
-                    FATEs = new HashSet<ushort>(from x in fates.Split(',') select ushort.Parse(x));
+                    FATEs = new HashSet<int>(from x in fates.Split(',') select int.Parse(x));
                 }
             }
         }
